test: split golden test
