+++ cmd/lx/golden_shared_test.go
@@ -568,291 +568,6 @@ func setupRelativePathsFixture(t *testing.T) (workDir, srcDir string) {
 	return dir, filepath.Join(dir, "src")
 }
 
-func TestGoldenWalk(t *testing.T) {
-	dir := setupWalkFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "001_walk_default", args: []string{"."}},
-		{name: "002_walk_compact", args: []string{"-n0", "."}},
-		{name: "003_specific_file", args: []string{"main.go"}},
-		{name: "004_multiple_roots", args: []string{"pkg", "README.md"}},
-		{name: "005_walk_subdir", args: []string{"src"}},
-		{name: "006_walk_parent_ref", args: []string{"src/../pkg"}},
-		{name: "007_walk_dot_slash", args: []string{"./src"}},
-	})
-}
-
-func TestGoldenFormatting(t *testing.T) {
-	dir := setupFormattingFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "010_fmt_xml", args: []string{"--xml", "."}},
-		{name: "011_fmt_html", args: []string{"--html", "main.go", "README.md"}},
-		{name: "012_fmt_markdown_explicit", args: []string{"--md", "main.go"}},
-		{name: "013_fmt_xml_sections", args: []string{"--xml", "-s", "Core", "main.go", "-s", "Docs", "README.md"}},
-		{name: "014_fmt_html_sections", args: []string{"--html", "-s", "Core", "main.go"}},
-		{name: "015_fmt_markdown_no_fences", args: []string{"-y", "configs/custom_template.yaml", "main.go"}},
-	})
-}
-
-func TestGoldenSections(t *testing.T) {
-	dir := setupSectionsFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "020_sections_explicit", args: []string{"-s", "Docs", "doc", "-s", "Source", "src"}},
-		{name: "021_prompts_mixed", args: []string{"-p", "Analyze this:", "main.go", "-p", "Determine bug"}},
-		{name: "022_xml_complex_structure", args: []string{"--xml", "-s", "Context", "-p", "Read carefully", "README.md", "-s", "Code", "main.go"}},
-		{name: "023_prompt_only", args: []string{"-p", "Just a question", "-p", "Another question"}},
-		{name: "024_section_empty", args: []string{"-s", "Empty Section"}},
-		{name: "025_interleaved_prompts_files", args: []string{"main.go", "-p", "Explain above", "pkg/util.go", "-p", "Explain below"}},
-	})
-}
-
-func TestGoldenFiltering(t *testing.T) {
-	dir := setupFilteringFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "030_include_only_go", args: []string{"-i", "*.go", "."}},
-		{name: "031_exclude_tests", args: []string{"-e", "*_test.go", "."}},
-		{name: "032_mixed_filters", args: []string{"-i", "*.go", "-e", "*_test.go", "."}},
-		{name: "033_filter_reset", args: []string{"-i", "*.md", ".", "-E", "-s", "All", "."}},
-		{name: "034_hidden_files", args: []string{"-H", "."}},
-		{name: "035_no_ignore_files", args: []string{"--no-ignore", "."}},
-		{name: "036_exclude_dir", args: []string{"-e", "src", "."}},
-		{name: "037_include_path_match", args: []string{"-i", "src/script.py", "."}},
-		{name: "038_nested_ignore_file", args: []string{"ignore_test"}},
-		{name: "039_complex_globs", args: []string{"-i", "**/*.{go,py}", "-e", "**/main*", "."}},
-	})
-}
-
-func TestGoldenSlicing(t *testing.T) {
-	dir := setupSlicingFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "040_lines_limit", args: []string{"--lines", "4", "."}},
-		{name: "041_line_numbers", args: []string{"-l", "main.go"}},
-		{name: "042_progressive_state", args: []string{
-			"-s", "Raw", "README.md",
-			"-s", "Numbered", "-l", "main.go",
-			"-s", "Sliced", "-n", "2", "pkg/util.go",
-			"-s", "Reset", "-L", "-N", "src/script.py",
-		}},
-		{name: "043_head_tail", args: []string{"--head", "3", "src/large.txt", "--tail", "2", "src/large.txt"}},
-		{name: "044_tail_only", args: []string{"--tail", "5", "src/large.txt"}},
-		{name: "045_head_only", args: []string{"--head", "5", "src/large.txt"}},
-		{name: "046_head_tail_overlap", args: []string{"--head", "100", "--tail", "100", "src/large.txt"}},
-		{name: "047_lines_0_compact", args: []string{"-n", "0", "main.go"}},
-		{name: "048_slice_binary", args: []string{"--head", "5", "bin/data.bin"}},
-	})
-}
-
-func TestGoldenSymlinks(t *testing.T) {
-	dir := setupSymlinksFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "050_symlinks_default", args: []string{"links"}},
-		{name: "051_symlinks_follow", args: []string{"--follow", "links"}},
-		{name: "052_symlinks_dag", args: []string{"--follow", "links/loop"}},
-		{name: "053_symlinks_infinite_cycle", args: []string{"--follow", "links/cycle_a"}},
-		{name: "054_file_links_explicit", args: []string{"--links", "links"}},
-		{name: "055_no_file_links", args: []string{"--no-links", "links"}},
-		{name: "056_follow_dirs_no_file_links", args: []string{"--follow", "--no-links", "links"}},
-		{name: "057_broken_symlink", args: []string{"links/broken_link"}},
-	})
-}
-
-func TestGoldenErrors(t *testing.T) {
-	dir := setupErrorsFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "060_access_denied_file", args: []string{"secret/locked.txt"}},
-		{name: "061_access_denied_dir", args: []string{"secret/locked_dir"}},
-		{name: "062_missing_file", args: []string{"nonexistent.go"}},
-		{name: "063_spaces_in_filename", args: []string{"spaces/file with spaces.txt"}},
-		{name: "064_spaces_in_dirname", args: []string{"spaces"}},
-		{name: "065_stdin_content", args: []string{"-"}, stdin: "Some content from pipe\nLine 2"},
-		{name: "066_stdin_and_files", args: []string{"main.go", "-"}, stdin: "Piped content"},
-		{name: "067_stdin_file_list", args: []string{}, stdin: "main.go\nREADME.md"},
-	})
-}
-
-func TestGoldenStats(t *testing.T) {
-	dir := setupStatsFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "070_stats_forced", args: []string{"--stats", "main.go"}},
-		{name: "071_quiet_mode", args: []string{"-q", "main.go"}},
-		{name: "072_verbose_debug", args: []string{"-vv", "main.go"}},
-		{name: "073_verbose_flag", args: []string{"--verbose=debug", "main.go"}},
-		{name: "074_no_stats", args: []string{"--no-stats", "main.go"}},
-	})
-}
-
-func TestGoldenDetection(t *testing.T) {
-	dir := setupDetectionFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "080_binary_detection", args: []string{"bin/data.bin"}},
-		{name: "081_empty_file", args: []string{"bin/empty.txt"}},
-		{name: "082_lang_detection_rs", args: []string{"langs/main.rs"}},
-		{name: "083_lang_detection_py", args: []string{"src/script.py"}},
-		{name: "084_lang_detection_dockerfile", args: []string{"langs/Dockerfile"}},
-		{name: "085_lang_detection_shebang", args: []string{"langs/script_no_ext"}},
-		{name: "086_render_image_md", args: []string{"assets/logo.png"}},
-		{name: "087_render_image_html", args: []string{"--html", "assets/logo.png"}},
-		{name: "088_render_image_compact", args: []string{"-n0", "assets/logo.png"}},
-	})
-}
-
-func TestGoldenConfig(t *testing.T) {
-	dir := setupConfigFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "090_config_hidden", args: []string{"-y", "configs/hidden.yaml", "."}},
-		{name: "091_config_follow", args: []string{"-y", "configs/follow.yaml", "links"}},
-		{name: "092_config_override", args: []string{"-y", "configs/follow.yaml", "--no-follow", "links"}},
-		{name: "093_force_file_flag", args: []string{"-f", ".gitignore"}},
-		{name: "094_force_file_flag_excludes", args: []string{"-e", "*.go", "-f", "main.go"}},
-		{name: "095_config_custom_sections", args: []string{"-y", "configs/custom_sections.yaml", "-s", "Head", "main.go"}},
-		{name: "096_direct_file_parent_ignore", args: []string{"direct_ignore_test/test.ignored", "direct_ignore_test/test.kept"}},
-	})
-}
-
-func TestGoldenComplex(t *testing.T) {
-	dir := setupComplexFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "100_complex_review_bundle", args: []string{
-			"--xml",
-			"-p", "Review this code",
-			"-s", "Configuration", "-i", "*.yaml", ".",
-			"-s", "Source", "-E", "-i", "*.go", "-e", "*_test.go", ".",
-			"-s", "Tests", "-E", "-i", "*_test.go", ".",
-		}},
-		{name: "101_stdin_null_terminated", args: []string{"-0"}, stdin: "main.go\x00README.md\x00spaces/file with spaces.txt\x00"},
-		{name: "102_walk_nested_respects_root_ignore", args: []string{"parent_ignore_test/level1/level2"}},
-		{name: "103_complex_ignore_exceptions", args: []string{"ignore_exception_test"}},
-	})
-}
-
-func TestGoldenArchive(t *testing.T) {
-	dir := setupArchiveFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "110_expand_archive_zip", args: []string{"-Z", "archive.zip"}},
-		{name: "111_expand_archive_no_flag", args: []string{"archive.zip"}},
-		{name: "112_expand_archive_filter", args: []string{"-Z", "-i", "*.go", "archive.zip"}},
-		{name: "113_expand_archive_config", args: []string{"-y", "configs/expand.yaml", "archive.zip"}},
-		{name: "114_expand_archive_no_expand_override", args: []string{"-y", "configs/expand.yaml", "--no-expand", "archive.zip"}},
-		{name: "115_expand_archive_tar_gz", args: []string{"-Z", "archive.tar.gz"}},
-	})
-}
-
-func TestGoldenDocuments(t *testing.T) {
-	dir := setupDocumentsFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "120_docs_pdf_extracted", args: []string{"sample.pdf"}},
-		{name: "121_docs_docx_extracted", args: []string{"sample.docx"}},
-		{name: "122_docs_xlsx_extracted", args: []string{"sample.xlsx"}},
-		{name: "123_docs_no_extract_flag", args: []string{"-D", "sample.pdf", "sample.docx", "sample.xlsx"}},
-		{name: "124_docs_odt_expanded", args: []string{"-Z", "sample.odt"}},
-		{name: "125_docs_odt_binary", args: []string{"sample.odt"}},
-		{name: "126_docs_pptx_extracted", args: []string{"sample.pptx"}},
-		{name: "127_docs_ods_expanded", args: []string{"-Z", "sample.ods"}},
-		{name: "128_docs_odp_expanded", args: []string{"-Z", "sample.odp"}},
-		{name: "129_docs_ods_binary", args: []string{"sample.ods"}},
-		{name: "130_docs_odp_binary", args: []string{"sample.odp"}},
-	})
-}
-
-func TestGoldenSkeleton(t *testing.T) {
-	dir := setupSkeletonFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "200_skeleton_go_functions", args: []string{"-F", "skeleton/main.go"}},
-		{name: "201_skeleton_go_structs", args: []string{"-T", "skeleton/main.go"}},
-		{name: "202_skeleton_go_both", args: []string{"-F", "-T", "skeleton/main.go"}},
-		{name: "203_skeleton_python_functions", args: []string{"-F", "skeleton/main.py"}},
-		{name: "204_skeleton_python_structs", args: []string{"-T", "skeleton/main.py"}},
-		{name: "205_skeleton_python_both", args: []string{"-F", "-T", "skeleton/main.py"}},
-		{name: "206_skeleton_c_functions", args: []string{"-F", "skeleton/main.c"}},
-		{name: "207_skeleton_c_structs", args: []string{"-T", "skeleton/main.c"}},
-		{name: "208_skeleton_c_both", args: []string{"-F", "-T", "skeleton/main.c"}},
-		{name: "209_skeleton_cpp_both", args: []string{"-F", "-T", "skeleton/main.cpp"}},
-		{name: "210_skeleton_java_structs", args: []string{"-T", "skeleton/Main.java"}},
-		{name: "211_skeleton_java_both", args: []string{"-F", "-T", "skeleton/Main.java"}},
-		{name: "212_skeleton_interleaved_reset", args: []string{"-F", "-T", "skeleton/main.py", "--reset-skeleton", "skeleton/main.py"}},
-		{name: "213_skeleton_toggle_off_functions", args: []string{"-F", "-T", "skeleton/main.go", "--no-functions", "skeleton/main.go"}},
-		{name: "214_skeleton_rust_both", args: []string{"-F", "-T", "skeleton/main.rs"}},
-		{name: "215_skeleton_typescript_both", args: []string{"-F", "-T", "skeleton/main.ts"}},
-		{name: "216_skeleton_kotlin_both", args: []string{"-F", "-T", "skeleton/main.kt"}},
-		{name: "217_skeleton_ruby_both", args: []string{"-F", "-T", "skeleton/main.rb"}},
-	})
-}
-
-func TestGoldenSkeletonSlicing(t *testing.T) {
-	dir := setupSkeletonFixture(t)
-	runTestGolden(t, dir, []goldenTestCase{
-		{name: "220_skeleton_lines_go_both", args: []string{"-F", "-T", "--lines", "3", "skeleton/main.go"}},
-		{name: "221_skeleton_head_go_both", args: []string{"-F", "-T", "--head", "2", "skeleton/main.go"}},
-		{name: "222_skeleton_tail_go_both", args: []string{"-F", "-T", "--tail", "2", "skeleton/main.go"}},
-		{name: "223_skeleton_head_tail_go_both", args: []string{"-F", "-T", "--head", "2", "--tail", "2", "skeleton/main.go"}},
-		{name: "224_skeleton_lines_python_both", args: []string{"-F", "-T", "--lines", "2", "skeleton/main.py"}},
-		{name: "225_skeleton_head_cpp_both", args: []string{"-F", "-T", "--head", "3", "skeleton/main.cpp"}},
-		{name: "226_skeleton_tail_java_both", args: []string{"-F", "-T", "--tail", "2", "skeleton/Main.java"}},
-		{name: "227_skeleton_lines_after_reset", args: []string{"-F", "-T", "--reset-skeleton", "--lines", "2", "skeleton/main.py"}},
-		{name: "228_skeleton_progressive_lines_head_tail", args: []string{"-F", "-T", "--lines", "3", "skeleton/main.go", "--head", "1", "skeleton/main.go", "--tail", "1", "skeleton/main.go"}},
-	})
-}
-
-func TestGoldenRelativePaths(t *testing.T) {
-	workDir, srcDir := setupRelativePathsFixture(t)
-
-	pkgDir, _ := os.Getwd()
-	t.Cleanup(func() { os.Chdir(pkgDir) })
-
-	canonicalWorkDir, _ := filepath.EvalSymlinks(workDir)
-
-	if err := os.Chdir(srcDir); err != nil {
-		t.Fatal(err)
-	}
-
-	perceivedRoot, _ := filepath.Abs("..")
-
-	cases := []goldenTestCase{
-		// --- 200-219: Basic Navigation ---
-		{name: "200_relative_parent_file", args: []string{"../README.md"}},
-		{name: "201_relative_sibling_file", args: []string{"../pkg/util.go"}},
-		{name: "202_relative_sibling_dir", args: []string{"../pkg"}},
-		{name: "203_mixed_dots_file", args: []string{"./../README.md"}},
-		{name: "204_parent_root_file", args: []string{"../.gitignore"}},
-		{name: "205_relative_symlink_target", args: []string{"../links/link_to_main.go"}},
-		{name: "206_zig_zag_file", args: []string{"../pkg/../README.md"}},
-		{name: "207_current_dir_explicit", args: []string{"./script.py"}},
-		{name: "208_redundant_relative", args: []string{"../src/script.py"}},
-		{name: "209_relative_binary", args: []string{"../bin/data.bin"}},
-		{name: "210_relative_access_denied", args: []string{"../secret/locked.txt"}},
-		{name: "211_relative_broken_link", args: []string{"../links/broken_link"}},
-
-		// --- 220-229: Formatting with Relative Paths ---
-		{name: "220_relative_xml", args: []string{"--xml", "../README.md"}},
-		{name: "221_relative_html_dir", args: []string{"--html", "../pkg"}},
-		{name: "222_relative_md_fences", args: []string{"-y", "../configs/custom_template.yaml", "../main.go"}},
-
-		// --- 230-239: Sections & Prompts ---
-		{name: "230_section_relative_parent", args: []string{"-s", "Root", "../README.md", "-s", "Lib", "../pkg"}},
-		{name: "231_interleaved_relative", args: []string{"../main.go", "-p", "Check this", "../pkg/util.go"}},
-		{name: "232_relative_config_loading", args: []string{"-y", "../configs/custom_sections.yaml", "-s", "Parent", "../README.md"}},
-
-		// --- 240-249: Filtering with Relative Roots ---
-		{name: "240_relative_include_globs", args: []string{"-i", "*.go", "../pkg"}},
-		{name: "241_relative_exclude_parent_dir", args: []string{"-e", "main*", ".."}},
-		{name: "242_relative_exclude_nested", args: []string{"-e", "*.txt", "../doc"}},
-		{name: "243_relative_hidden_override", args: []string{"-H", "../.hidden"}},
-		{name: "244_relative_force_include", args: []string{"-f", "../.gitignore"}},
-
-		// --- 250-259: Slicing with Relative Paths ---
-		{name: "250_relative_lines_limit", args: []string{"-n", "2", "../main.go"}},
-		{name: "251_relative_head_tail", args: []string{"--head", "2", "--tail", "2", "../src/large.txt"}},
-		{name: "252_relative_line_numbers", args: []string{"-l", "../pkg/util.go"}},
-		{name: "253_relative_compact", args: []string{"-n", "0", "../README.md"}},
-
-		// --- 260-269: Stdin & Complex ---
-		{name: "260_stdin_relative_paths", args: []string{"-"}, stdin: "../README.md\n../pkg/util.go"},
-		{name: "261_stdin_relative_null", args: []string{"-0"}, stdin: "../main.go\x00../bin/data.bin\x00"},
-		{name: "262_symlink_follow_relative", args: []string{"--follow", "../links/loop"}},
-	}
-
-	runGoldenTests(t, cases, pkgDir, workDir, canonicalWorkDir, perceivedRoot)
-}
-
 func runGoldenTests(t *testing.T, cases []goldenTestCase, pkgDir string, scrubPaths ...string) {
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
